pyproject.toml
[tool.poetry]
name = "finaltest"
version = "0.1.0"
description = ""
authors = ["suekwon <sue11@snu.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
